Add rfp ordering by status so open rfps shown first
This is jank because we have a third status that
makes time-dependent statuses based on an open and
close date which happens to be after CLOSED in the
enum stored in the database (used in prisma sorting).

Sorting post-fetch is also jank though because it
can only sort within the paginated result from the db,
meaning we would have to fetch all rfps and manually
manage take size and pagination post-sorting.